// DrTech.Web/src/app/pages/ngo/dto/need-dto.ts
export class NeedDTO {
  id: string;
  fileName: string;
  type: number;
  typeDescription: string;
  subType: number;
  subTypeDescription: string;
  description: string;
  userId: string;
  donateTo: number;
  deliveryDate: string;
  status: number;
  statusDesciption: string;
  city: number;
  cityDescription: string;
}